Fixes clang9 warnings.

+++ src/animation/offline/animation_builder.cc
@@ -163,7 +163,7 @@ void CompressQuat(const ozz::math::Quaternion& _src,
                   ozz::animation::QuaternionKey* _dest) {
   // Finds the largest quaternion component.
   const float quat[4] = {_src.x, _src.y, _src.z, _src.w};
-  const size_t largest = std::max_element(quat, quat + 4, LessAbs) - quat;
+  const ptrdiff_t largest = std::max_element(quat, quat + 4, LessAbs) - quat;
   assert(largest <= 3);
   _dest->largest = largest & 0x3;
 
@@ -269,7 +269,7 @@ unique_ptr<Animation> AnimationBuilder::operator()(
 
   // Declares and preallocates tracks to sort.
   size_t translations = 0, rotations = 0, scales = 0;
-  for (int i = 0; i < num_tracks; ++i) {
+  for (uint16_t i = 0; i < num_tracks; ++i) {
     const RawAnimation::JointTrack& raw_track = _input.tracks[i];
     translations += raw_track.translations.size() + 2;  // +2 because worst case
     rotations += raw_track.rotations.size() + 2;        // needs to add the
